fix(заявки): telegram_id заявки был 32-битным — падал INSERT на новых Telegram id
leads.telegram_id хранился как INTEGER, а Telegram уже выдаёт id за пределами
этого диапазона (у владельца — 7678374811). Любая заявка — корзина с
промокодом, price_offer, сценарная — падала на INSERT с "integer out of
range", и человек видел общее "internal error" без объяснения.

users.telegram_id и notifications.chat_id уже были BigInteger; leads —
единственное исключение. Правится и модель, и мини-миграция существующей
колонки (ALTER COLUMN ... TYPE BIGINT).

Воспроизведено на проде (реальная ошибка в логах backend), пофикшено и
проверено там же: тот же INSERT, который падал, теперь проходит. Тест
проверяет объявленный тип колонки — SQLite границы типов не соблюдает и
round-trip-тест на нём молчал бы даже с багом.

+++ backend/app/models/lead.py
@@ -6,7 +6,7 @@
 """
 from datetime import datetime
 
-from sqlalchemy import JSON, DateTime, Index, Integer, Numeric, String, Text, func, text
+from sqlalchemy import JSON, BigInteger, DateTime, Index, Integer, Numeric, String, Text, func, text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.session import Base
@@ -54,7 +54,9 @@ class Lead(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True)
     user_id: Mapped[int | None] = mapped_column(Integer, index=True)
-    telegram_id: Mapped[int | None] = mapped_column(Integer, index=True)
+    # BigInteger, как у users.telegram_id: Telegram уже выдаёт id за пределами
+    # 32-битного INTEGER (например 7678374811) — заявка падала на INSERT.
+    telegram_id: Mapped[int | None] = mapped_column(BigInteger, index=True)
     name: Mapped[str | None] = mapped_column(String(200))
     phone: Mapped[str | None] = mapped_column(String(64))
     username: Mapped[str | None] = mapped_column(String(200))
